Create header
The header class was displayed as a flex containes, and the ul tag was
also displayed as a flex container. To the former some properties were
added, such as justify-content: space-between and align-items: center. To
the latter the margin and the padding were set to 0 and a gap rule was
addeds to separate the items inside it.