Fix the modal layout
By grouping in another div all the elements except for the icon, the
flex properties would organize most of the layout. Having in
consideration that the icon would shrink, its flex-shrink property was
set to 0. The close icon was grouped inside the header class and with
some tuning the desired outcome was accomplished.