Create the Holy Grail layout
Added a div wrapping the cards and the sidebar. Made the body display
flex and as a column. For the header and the footer, made flex 0 and
added some padding to achieve the behavior. This could have been also
done by treating both as flex containers. for the middle content, added
a class and displayed as row with flex 1 to push the footer to the
bottom.